alpha v.0.2

+++ src/main/java/fantasticfour/controllers/dao/ProductDao.java
@@ -3,5 +3,5 @@ package fantasticfour.controllers.dao;
 import fantasticfour.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductDao extends CrudRepository<Product, Long> {
+public interface ProductDao extends CrudRepository<Product, Integer> {
 }
